integrated train and recog functions into a single script
