+++ src/bin/asr_recog.py
@@ -6,25 +6,11 @@
 
 
 import argparse
-import json
 import logging
 import os
-import pickle
 import random
 
-# chainer related
-import chainer
 import numpy as np
-
-# spnet related
-from e2e_asr_attctc import E2E
-from e2e_asr_attctc import Loss
-
-# for kaldi io
-import kaldi_io_py
-
-# rnnlm
-import lm_train
 
 
 def main():
@@ -85,70 +71,22 @@ def main():
     # display PYTHONPATH
     logging.info('python path = ' + os.environ['PYTHONPATH'])
 
-    # display chainer version
-    logging.info('chainer version = ' + chainer.__version__)
-
     # seed setting (chainer seed may not need it)
     nseed = args.seed
     random.seed(nseed)
     np.random.seed(nseed)
-    os.environ["CHAINER_SEED"] = str(nseed)
-    logging.info('chainer seed = ' + os.environ['CHAINER_SEED'])
+    logging.info('set random seed = %d' % nseed)
 
-    # read training config
-    with open(args.model_conf, "rb") as f:
-        logging.info('reading a model config file from' + args.model_conf)
-        idim, odim, train_args = pickle.load(f)
-
-    for key in sorted(vars(args).keys()):
-        logging.info('ARGS: ' + key + ': ' + str(vars(args)[key]))
-
-    # specify model architecture
-    logging.info('reading model parameters from' + args.model)
-    e2e = E2E(idim, odim, train_args)
-    model = Loss(e2e, train_args.mtlalpha)
-    chainer.serializers.load_npz(args.model, model)
-
-    # read rnnlm
-    if args.rnnlm:
-        rnnlm = lm_train.ClassifierWithState(lm_train.RNNLM(len(train_args.char_list), 650))
-        chainer.serializers.load_npz(args.rnnlm, rnnlm)
+    # recog
+    logging.info('backend = ' + args.backend)
+    if args.backend == "chainer":
+        from asr_chainer import recog
+        recog(args)
+    elif args.backend == "pytorch":
+        from asr_pytorch import recog
+        recog(args)
     else:
-        rnnlm = None
-
-    # prepare Kaldi reader
-    reader = kaldi_io_py.read_mat_ark(args.recog_feat)
-
-    # read json data
-    with open(args.recog_label, 'rb') as f:
-        recog_json = json.load(f)['utts']
-
-    new_json = {}
-    for name, feat in reader:
-        logging.info('decoding ' + name)
-        y_hat = e2e.recognize(feat, args, train_args.char_list, rnnlm)
-        y_true = map(int, recog_json[name]['tokenid'].split())
-
-        # print out decoding result
-        seq_hat = [train_args.char_list[int(idx)] for idx in y_hat]
-        seq_true = [train_args.char_list[int(idx)] for idx in y_true]
-        seq_hat_text = "".join(seq_hat).replace('<space>', ' ')
-        seq_true_text = "".join(seq_true).replace('<space>', ' ')
-        logging.info("groundtruth[%s]: " + seq_true_text, name)
-        logging.info("prediction [%s]: " + seq_hat_text, name)
-
-        # copy old json info
-        new_json[name] = recog_json[name]
-
-        # added recognition results to json
-        new_json[name]['rec_tokenid'] = " ".join(
-            [str(idx[0]) for idx in y_hat])
-        new_json[name]['rec_token'] = " ".join(seq_hat)
-        new_json[name]['rec_text'] = seq_hat_text
-
-    # TODO(watanabe) fix character coding problems when saving it
-    with open(args.result_label, 'wb') as f:
-        f.write(json.dumps({'utts': new_json}, indent=4).encode('utf_8'))
+        raise ValueError("chainer and pytorch are only supported.")
 
 
 if __name__ == '__main__':
